tryingfloat32 to stabilize training

+++ scripts/input_to_label_and_rationale.py
@@ -318,7 +318,7 @@ def main():
     logger.info("Loading pretrained tokenizer...")
 
     ### Change model to llama (make this more dynamic like t5 and gpt3, remove token)
-    tokenizer = LlamaTokenizer.from_pretrained("meta-llama/Llama-2-7b-hf", token=os.environ['HF_TOKEN'], pad_token = '[PAD]', torch_dtype=torch.float16)
+    tokenizer = LlamaTokenizer.from_pretrained("meta-llama/Llama-2-7b-hf", token=os.environ['HF_TOKEN'], pad_token = '[PAD]', torch_dtype=torch.float32)
     # model = AutoModelForCausalLM.from_pretrained("meta-llama/Llama-2-7b-hf", token=os.environ['HF_TOKEN'])
     # tokenizer = tokenizer_name.from_pretrained(model_args.tokenizer_name)#, cache_dir=model_args.cache_dir)
     #print("tokenizer for model loaded successfully")
@@ -326,7 +326,7 @@ def main():
         model_name = MODEL_MAPPING[model_class]
         if model_args.pretrained_model_file:
             # model = T5ForConditionalGeneration.from_pretrained(model_args.pretrained_model_file)
-            model = LlamaForCausalLM.from_pretrained("meta-llama/Llama-2-7b-hf", token=os.environ['HF_TOKEN'], torch_dtype=torch.float16)
+            model = LlamaForCausalLM.from_pretrained("meta-llama/Llama-2-7b-hf", token=os.environ['HF_TOKEN'], torch_dtype=torch.float32)
 
             if model_args.dropout_rate:
                 raise Exception("can't update/specify dropout currently when load pretrained model from directory")
@@ -336,10 +336,10 @@ def main():
             logger.info("Loading pretrained model")
             if model_args.dropout_rate:
                 # model = model_name.from_pretrained(model_args.model_type, dropout_rate=model_args.dropout_rate)
-                model = LlamaForCausalLM.from_pretrained("meta-llama/Llama-2-7b-hf", token=os.environ['HF_TOKEN'], torch_dtype=torch.float16, dropout_rate=model_args.dropout_rate)
+                model = LlamaForCausalLM.from_pretrained("meta-llama/Llama-2-7b-hf", token=os.environ['HF_TOKEN'], torch_dtype=torch.float32, dropout_rate=model_args.dropout_rate)
             else:
                 # model = model_name.from_pretrained(model_args.model_type)
-                model = LlamaForCausalLM.from_pretrained("meta-llama/Llama-2-7b-hf", token=os.environ['HF_TOKEN'], torch_dtype=torch.float16)
+                model = LlamaForCausalLM.from_pretrained("meta-llama/Llama-2-7b-hf", token=os.environ['HF_TOKEN'], torch_dtype=torch.float32)
         else:
             # load model from scratch with no pretrained weights
             config_name = CONFIG_MAPPING[model_class]()
